Split the generally very large orblib(box).dat into orblib(box)_qgrid.dat and orblib(box)_losvd_hist.dat. When reading the orblib, both variants (one file or two files) are supported for compatibility.

+++ dynamite/orblib.py
@@ -89,10 +89,18 @@ class LegacyOrbitLibrary(OrbitLibrary):
             - +8 log and status files
 
         """
-        # check if orbit library was calculated already
+        # check if orbit library was calculated already (FIXME: improve this!)
         check1 = os.path.isfile(self.mod_dir+'datfil/orblib.dat.bz2')
         check2 = os.path.isfile(self.mod_dir+'datfil/orblibbox.dat.bz2')
-        if not check1 or not check2:
+        if check1 and check2:
+            calculate_orblib = False
+        else:
+            check1 = os.path.isfile(self.mod_dir+'datfil/orblib_qgrid.dat.bz2')
+            check1 = check1 and os.path.isfile(self.mod_dir+'datfil/orblib_losvd_hist.dat.bz2')
+            check1 = check1 and os.path.isfile(self.mod_dir+'datfil/orblibbox_qgrid.dat.bz2')
+            check1 = check1 and os.path.isfile(self.mod_dir+'datfil/orblibbox_losvd_hist.dat.bz2')
+            calculate_orblib = False if check1 else True
+        if calculate_orblib:
             # prepare the fortran input files for orblib
             self.create_fortran_input_orblib(self.mod_dir+'infil/')
             if self.system.is_bar_disk_system():
@@ -363,18 +371,31 @@ class LegacyOrbitLibrary(OrbitLibrary):
                 label = f'[binfile for populations aperture {1+i}]'
                 line = f'"infil/{pop_i.binfile}"{tab[:-3]}{label}\n'
                 f.write(line)
-            if box:
-                f.write(f'datfil/orblibbox.dat\n')
+            o_file = 'datfil/orblibbox' if box else 'datfil/orblib'
+            if self.system.is_bar_disk_system() or self.settings['use_new_mirroring']:
+                f_name = f'"{o_file}_qgrid.dat"'
+                f.write(f'{f_name}{tab[:-4] if len(f_name) >= 32 else tab[:-3]}'
+                        '[orbit qgrid file]\n')
+                f_name = f'"{o_file}_losvd_hist.dat"'
+                f.write(f'{f_name}{tab[:-4] if len(f_name) >= 32 else tab[:-3]}'
+                        '[orbit losvd 1d hist file]\n')
+                f_name = f'"{o_file}.dat_orbclass.out"'
+                f.write(f'{f_name}{tab[:-4] if len(f_name) >= 32 else tab[:-3]}'
+                        '[orbit classification file]\n')
             else:
-                f.write(f'datfil/orblib.dat\n')
+                f.write(f'{o_file}.dat\n')
             f.close()
         write_orblib_dot_in(box=False)
         write_orblib_dot_in(box=True)
         #-------------------
         #write triaxmass.in
         #-------------------
+        if self.system.is_bar_disk_system() or self.settings['use_new_mirroring']:
+            orblib_file = 'orblib_qgrid.dat'
+        else:
+            orblib_file = 'orblib.dat'
         text='infil/parameters_lum.in' +'\n' + \
-        'datfil/orblib.dat' +'\n' + \
+        f'datfil/{orblib_file}' +'\n' + \
         'datfil/mass_radmass.dat' +'\n' + \
         'datfil/mass_qgrid.dat'
         triaxmass_file= open(path+'triaxmass.in',"w")
@@ -569,7 +590,12 @@ class LegacyOrbitLibrary(OrbitLibrary):
             txt_file.write(f'test -e {self.legacy_directory}/{f_name} || ' +
                            f'{{ echo "File {self.legacy_directory}/{f_name} ' +
                            'not found." && exit 127; }\n')
-        txt_file.write('(rm -f datfil/orblib.dat.tmp datfil/orblib.dat\n')
+        if self.system.is_bar_disk_system() or self.settings['use_new_mirroring']:
+            txt_file.write('(rm -f datfil/orblib.dat.tmp '
+                           'datfil/orblib_qgrid.dat '
+                           'datfil/orblib_losvd_hist.dat\n')
+        else:
+            txt_file.write('(rm -f datfil/orblib.dat.tmp datfil/orblib.dat\n')
         txt_file.write(f'{self.legacy_directory}/{orb_prgrm} < infil/orblib.in '
                         '>> datfil/orblib.log\n')
         txt_file.write('rm -f datfil/mass_qgrid.dat datfil/mass_radmass.dat '
@@ -586,16 +612,38 @@ class LegacyOrbitLibrary(OrbitLibrary):
             txt_file.write(f'{self.legacy_directory}/triaxmassbin '
                            '< infil/triaxmassbin.in >> datfil/triaxmassbin.log\n')
 
-        txt_file.write('rm -f datfil/orblib.dat.bz2 '
-                        '&& bzip2 -k datfil/orblib.dat\n')
-        txt_file.write('rm datfil/orblib.dat) &\n')
+        if self.system.is_bar_disk_system() or self.settings['use_new_mirroring']:
+            txt_file.write('rm -f datfil/orblib_qgrid.dat.bz2 '
+                        '&& bzip2 -k datfil/orblib_qgrid.dat\n')
+            txt_file.write('rm datfil/orblib_qgrid.dat\n')
+            txt_file.write('rm -f datfil/orblib_losvd_hist.dat.bz2 '
+                        '&& bzip2 -k datfil/orblib_losvd_hist.dat\n')
+            txt_file.write('rm datfil/orblib_losvd_hist.dat) &\n')
+        else:
+            txt_file.write('rm -f datfil/orblib.dat.bz2 '
+                            '&& bzip2 -k datfil/orblib.dat\n')
+            txt_file.write('rm datfil/orblib.dat) &\n')
         txt_file.write('orblib=$!\n')
-        txt_file.write('(rm -f datfil/orblibbox.dat.tmp datfil/orblibbox.dat\n')
+        if self.system.is_bar_disk_system() or self.settings['use_new_mirroring']:
+            txt_file.write('(rm -f datfil/orblibbox.dat.tmp '
+                           'datfil/orblibbox_qgrid.dat '
+                           'datfil/orblibbox_losvd_hist.dat\n')
+        else:
+            txt_file.write('(rm -f datfil/orblibbox.dat.tmp '
+                           'datfil/orblibbox.dat\n')
         txt_file.write(f'{self.legacy_directory}/{orb_prgrm} '
                        '< infil/orblibbox.in >> datfil/orblibbox.log\n')
-        txt_file.write('rm -f datfil/orblibbox.dat.bz2 '
-                       '&& bzip2 -k datfil/orblibbox.dat\n')
-        txt_file.write('rm datfil/orblibbox.dat) &\n')
+        if self.system.is_bar_disk_system() or self.settings['use_new_mirroring']:
+            txt_file.write('rm -f datfil/orblibbox_qgrid.dat.bz2 '
+                        '&& bzip2 -k datfil/orblibbox_qgrid.dat\n')
+            txt_file.write('rm datfil/orblibbox_qgrid.dat\n')
+            txt_file.write('rm -f datfil/orblibbox_losvd_hist.dat.bz2 '
+                        '&& bzip2 -k datfil/orblibbox_losvd_hist.dat\n')
+            txt_file.write('rm datfil/orblibbox_losvd_hist.dat) &\n')
+        else:
+            txt_file.write('rm -f datfil/orblibbox.dat.bz2 '
+                        '&& bzip2 -k datfil/orblibbox.dat\n')
+            txt_file.write('rm datfil/orblibbox.dat) &\n')
         txt_file.write('orblibbox=$!\n')
         txt_file.write('wait $orblib $orblibbox\n')
         txt_file.close()
@@ -620,8 +668,15 @@ class LegacyOrbitLibrary(OrbitLibrary):
             txt_file.write(f'test -e {self.legacy_directory}/{f_name} || ' +
                            f'{{ echo "File {self.legacy_directory}/{f_name} ' +
                            'not found." && exit 127; }\n')
-        txt_file.write('rm -f datfil/orblib.dat.tmp datfil/orblib.dat '
-                       'datfil/orblib.dat.bz2\n')
+        if self.system.is_bar_disk_system() or self.settings['use_new_mirroring']:
+            txt_file.write('rm -f datfil/orblib.dat.tmp '
+                           'datfil/orblib_qgrid.dat '
+                           'datfil/orblib_qgrid.dat.bz2 '
+                           'datfil/orblib_losvd_hist.dat '
+                           'datfil/orblib_losvd_hist.dat.bz2\n')
+        else:
+            txt_file.write('rm -f datfil/orblib.dat.tmp datfil/orblib.dat '
+                           'datfil/orblib.dat.bz2\n')
         txt_file.write(f'{self.legacy_directory}/{orb_prgrm} < infil/orblib.in '
                        '>> datfil/orblib.log\n')
         txt_file.write('rm -f datfil/mass_qgrid.dat datfil/mass_radmass.dat '
@@ -630,10 +685,17 @@ class LegacyOrbitLibrary(OrbitLibrary):
                        '< infil/triaxmass.in >> datfil/triaxmass.log\n')
         txt_file.write(f'{self.legacy_directory}/triaxmassbin '
                        '< infil/triaxmassbin.in >> datfil/triaxmassbin.log\n')
-        txt_file.write(
-            'bzip2 -kc datfil/orblib.dat > datfil/orblib.dat.staging.bz2 '
-            '&& mv datfil/orblib.dat.staging.bz2 datfil/orblib.dat.bz2\n')
-        txt_file.write('rm datfil/orblib.dat\n')
+        if self.system.is_bar_disk_system() or self.settings['use_new_mirroring']:
+            for f_name in 'datfil/orblib_qgrid', 'datfil/orblib_losvd_hist':
+                txt_file.write(
+                    f'bzip2 -kc {f_name}.dat > {f_name}.dat.staging.bz2 '
+                    f'&& mv {f_name}.dat.staging.bz2 {f_name}.dat.bz2\n')
+                txt_file.write(f'rm {f_name}.dat\n')
+        else:
+            txt_file.write(
+                'bzip2 -kc datfil/orblib.dat > datfil/orblib.dat.staging.bz2 '
+                '&& mv datfil/orblib.dat.staging.bz2 datfil/orblib.dat.bz2\n')
+            txt_file.write('rm datfil/orblib.dat\n')
         txt_file.close()
         # boxorbits
         cmdstr_box = 'cmd_box_orbs'
@@ -643,14 +705,28 @@ class LegacyOrbitLibrary(OrbitLibrary):
         txt_file.write(f'test -e {self.legacy_directory}/{orb_prgrm} || ' +
                        f'{{ echo "File {self.legacy_directory}/{orb_prgrm} ' +
                        'not found." && exit 127; }\n')
-        txt_file.write('rm -f datfil/orblibbox.dat.tmp datfil/orblibbox.dat '
-                       'datfil/orblibbox.dat.bz2\n')
+        if self.system.is_bar_disk_system() or self.settings['use_new_mirroring']:
+            txt_file.write('rm -f datfil/orblibbox.dat.tmp '
+                           'datfil/orblibbox_qgrid.dat '
+                           'datfil/orblibbox_qgrid.dat.bz2 '
+                           'datfil/orblibbox_losvd_hist.dat '
+                           'datfil/orblibbox_losvd_hist.dat.bz2\n')
+        else:
+            txt_file.write('rm -f datfil/orblibbox.dat.tmp datfil/orblibbox.dat '
+                        'datfil/orblibbox.dat.bz2\n')
         txt_file.write(f'{self.legacy_directory}/{orb_prgrm} '
                        '< infil/orblibbox.in >> datfil/orblibbox.log\n')
-        txt_file.write(
-            'bzip2 -kc datfil/orblibbox.dat > datfil/orblibbox.dat.staging.bz2 '
-            '&& mv datfil/orblibbox.dat.staging.bz2 datfil/orblibbox.dat.bz2\n')
-        txt_file.write('rm datfil/orblibbox.dat\n')
+        if self.system.is_bar_disk_system() or self.settings['use_new_mirroring']:
+            for f_name in 'datfil/orblibbox_qgrid', 'datfil/orblibbox_losvd_hist':
+                txt_file.write(
+                    f'bzip2 -kc {f_name}.dat > {f_name}.dat.staging.bz2 '
+                    f'&& mv {f_name}.dat.staging.bz2 {f_name}.dat.bz2\n')
+                txt_file.write(f'rm {f_name}.dat\n')
+        else:
+            txt_file.write(
+                'bzip2 -kc datfil/orblibbox.dat > datfil/orblibbox.dat.staging.bz2 '
+                '&& mv datfil/orblibbox.dat.staging.bz2 datfil/orblibbox.dat.bz2\n')
+            txt_file.write('rm datfil/orblibbox.dat\n')
         txt_file.close()
         # returns the name of the executables
         return cmdstr_tube, cmdstr_box
@@ -662,6 +738,12 @@ class LegacyOrbitLibrary(OrbitLibrary):
     def read_orbit_base(self, fileroot, return_instrisic_moments=False):
         """
         Read orbit library from file datfil/{fileroot}.dat.bz2'
+
+        Depending on the DYNAMITE version, the orbit library will reside in
+        either datfil/{fileroot}.dat.bz2 (legacy behavior) or in two separate
+        files datfil/{fileroot}_qgrid.dat.bz2 and
+        datfil/{fileroot}_losvd_hist.dat.bz2 (new behavior). If both
+        alterntives exist, default to the new behavior.
 
         Parameters
         ----------
@@ -683,27 +765,64 @@ class LegacyOrbitLibrary(OrbitLibrary):
         """
         cur_dir = os.getcwd()
         os.chdir(self.mod_dir)
+        check = os.path.isfile(f'datfil/{fileroot}_qgrid.dat.bz2')
+        check = check and os.path.isfile(f'datfil/{fileroot}_losvd_hist.dat.bz2')
+        legacy_file = False if check else True
         # unzip orblib to a temporary file with ml value attached
         # ml value is needed to prevent different processes clashing
         ml = self.parset['ml']
-        tmpfname = f'datfil/{fileroot}_{ml}.dat'
-        subprocess.run(f'bunzip2 -c datfil/{fileroot}.dat.bz2 > {tmpfname}',
-                       shell=True)
+        if legacy_file:
+            orblib_file = f'datfil/{fileroot}.dat.bz2'
+            tmpfname = f'datfil/{fileroot}_{ml}.dat'
+        else:
+            orblib_file = f'datfil/{fileroot}_qgrid.dat.bz2'
+            tmpfname = f'datfil/{fileroot}_qgrid_{ml}.dat'
+        subprocess.run(f'bunzip2 -c {orblib_file} > {tmpfname}', shell=True)
         # read the fortran file
-        orblibf = FortranFile(tmpfname, 'r')
+        orblib_in = FortranFile(tmpfname, 'r')
         # read size of orbit library
         # from integrator_setup_write, lines 506 - 5129:
-        norb, _, _, _, ndith = orblibf.read_ints(np.int32)
+        norb, _, _, _, ndith = orblib_in.read_ints(np.int32)
         # from qgrid_setup_write, lines 2339-1350:
-        quad_light_grid_sizes = orblibf.read_ints(np.int32)
+        quad_light_grid_sizes = orblib_in.read_ints(np.int32)
         size_ql, size_qph, size_qth, size_qlr = quad_light_grid_sizes
-        quad_lr = orblibf.read_reals(float)
-        quad_lth = orblibf.read_reals(float)
-        quad_lph = orblibf.read_reals(float)
+        quad_lr = orblib_in.read_reals(float)
+        quad_lth = orblib_in.read_reals(float)
+        quad_lph = orblib_in.read_reals(float)
         if return_instrisic_moments:
             instrisic_grid = [quad_lr, quad_lth, quad_lph]
+            intrinsic_moms = np.zeros((norb, size_qlr, size_qth, size_qph, 16))
+        orbtypes = np.zeros((norb, ndith**3), dtype=int)
+        density_3D = np.zeros((norb, size_qlr, size_qth, size_qph))
+        if not legacy_file:  # Enter this loop if orblib is in multiple files
+            for j in range(norb):
+                _, _, _, _, _ = orblib_in.read_ints(np.int32)
+                orbtypes[j, :] = orblib_in.read_ints(np.int32)
+                quad_light = orblib_in.read_reals(float)
+                quad_light = np.reshape(quad_light, quad_light_grid_sizes[::-1])
+                # quad_light stores orbit features in 3D (r,th,phi) bins.
+                # Quad_light[ir,it,ip,XXX] stores
+                # - the zeroth moment i.e. density for XXX=0,
+                # - the first moments x,y,z,vx,vy,vz for XXX=slice(1,7)
+                # - 2nd moments vx^2,vy^2,vz^2,vx*vy,vy*vz,vz*vx for XXX=slice(7,13)
+                # - an averaged orbit classification for XXX=slice(13,None)
+                # in the bin indexed by (ir,it,ip).
+                # We need to extract 3D density for use in weight solving.
+                density_3D[j] = quad_light[:,:,:,0]
+                if return_instrisic_moments:
+                    intrinsic_moms[j] = quad_light
+            # done with orblib_qgrid.dat.bz2,
+            # now switch to orblib_losvd_hist.dat.bz2
+            orblib_in.close()
+            # remove temporary file
+            os.remove(tmpfname)
+            orblib_file = f'datfil/{fileroot}_losvd_hist.dat.bz2'
+            tmpfname = f'datfil/{fileroot}_losvd_hist_{ml}.dat'
+            subprocess.run(f'bunzip2 -c {orblib_file} > {tmpfname}', shell=True)
+            # read the fortran file
+            orblib_in = FortranFile(tmpfname, 'r')
         # from histogram_setup_write, lines 1917-1926:
-        tmp = orblibf.read_record(np.int32, np.int32, float)
+        tmp = orblib_in.read_record(np.int32, np.int32, float)
         nconstr = tmp[0][0] # = total number of apertures for ALL kinematics
                             #   and populations with own apertures
         # nvhist = tmp[1][0] # = (nvbins-1)/2 for histo of FIRST kinematic set
@@ -748,37 +867,34 @@ class LegacyOrbitLibrary(OrbitLibrary):
         tmp = zip(hist_bins,n_apertures)
         velhist0 = [np.zeros((norb, nv, na)) for (nv,na) in tmp]
         # Next read the histograms themselves.
-        orbtypes = np.zeros((norb, ndith**3), dtype=int)
-        density_3D = np.zeros((norb, size_qlr, size_qth, size_qph))
-        if return_instrisic_moments:
-            intrinsic_moms = np.zeros((norb, size_qlr, size_qth, size_qph, 16))
         for j in range(norb):
-            _, _, _, _, _ = orblibf.read_ints(np.int32)
-            orbtypes[j, :] = orblibf.read_ints(np.int32)
-            quad_light = orblibf.read_reals(float)
-            quad_light = np.reshape(quad_light, quad_light_grid_sizes[::-1])
-            # quad_light stores orbit features in 3D (r,th,phi) bins.
-            # Quad_light[ir,it,ip,XXX] stores
-            # - the zeroth moment i.e. density for XXX=0,
-            # - the first moments x,y,z,vx,vy,vz for XXX=slice(1,7)
-            # - 2nd moments vx^2,vy^2,vz^2,vx*vy,vy*vz,vz*vx for XXX=slice(7,13)
-            # - an averaged orbit classification for XXX=slice(13,None)
-            # in the bin indexed by (ir,it,ip).
-            # We need to extract 3D density for use in weight solving.
-            density_3D[j] = quad_light[:,:,:,0]
+            if legacy_file:  # Enter this loop if orblib is in a single file
+                _, _, _, _, _ = orblib_in.read_ints(np.int32)
+                orbtypes[j, :] = orblib_in.read_ints(np.int32)
+                quad_light = orblib_in.read_reals(float)
+                quad_light = np.reshape(quad_light, quad_light_grid_sizes[::-1])
+                # quad_light stores orbit features in 3D (r,th,phi) bins.
+                # Quad_light[ir,it,ip,XXX] stores
+                # - the zeroth moment i.e. density for XXX=0,
+                # - the first moments x,y,z,vx,vy,vz for XXX=slice(1,7)
+                # - 2nd moments vx^2,vy^2,vz^2,vx*vy,vy*vz,vz*vx for XXX=slice(7,13)
+                # - an averaged orbit classification for XXX=slice(13,None)
+                # in the bin indexed by (ir,it,ip).
+                # We need to extract 3D density for use in weight solving.
+                density_3D[j] = quad_light[:,:,:,0]
+                if return_instrisic_moments:
+                    intrinsic_moms[j] = quad_light
             for i_ap in range(nconstr):
                 kinpop_idx = kinpop_idx_per_ap[i_ap]
                 i_ap0 = i_ap - idx_ap_reset[kinpop_idx]
-                ivmin, ivmax = orblibf.read_ints(np.int32)
+                ivmin, ivmax = orblib_in.read_ints(np.int32)
                 if ivmin <= ivmax:
                     nv0 = (hist_bins[kinpop_idx]-1)/2
                     # ^--- this is an integer since hist_bins is odd
                     nv0 = int(nv0)
-                    tmp = orblibf.read_reals(float)
+                    tmp = orblib_in.read_reals(float)
                     velhist0[kinpop_idx][j, ivmin+nv0:ivmax+nv0+1, i_ap0] = tmp
-            if return_instrisic_moments:
-                intrinsic_moms[j] = quad_light
-        orblibf.close()
+        orblib_in.close()
         # remove temporary file
         os.remove(tmpfname)
         os.chdir(cur_dir)
